Added delete functionality
Added a render function that's more react-like in that it takes the model
and completely redoes the dom. No shadow dom comparisons though :). There's
still some two way data binding, but we'd need to make dispatch
functionality for the model. Also the attempt to use Proxy failed as it
doesn't do deep comparisons.